Commit 002: changed to accept double values instead of int for extensibility; passing analysis arguments as pointers now; wrote separate function to check if shape is a triangle; replaced printf_s with puts/printf for universality

+++ PolygonChecker/main.h
@@ -5,4 +5,4 @@
 
 void printWelcome();
 int printShapeMenu();
-int* getTriangleSides(int* triangleSides);
+double* getTriangleSides(double* triangleSides);
